Refactors node writing and reading functions to take BitString{Reader,Writer}s
This fixes a bug that I didn't have a test for yet where there would be a gap
between the tree and contents. Strange the things you think of when you're
trying to fall asleep. I didn't want to have to implement moving forward to the
next bit in the reader. So we get to save a byte instead.

```
wc -c wild-wild-west*.huff
2563 wild-wild-west-no-gap.huff
2564 wild-wild-west-with-gap.huff
```